models/ml_models.py
import numpy as np
from typing import Dict, List, Optional
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
import joblib

class MLPredictor:
    """机器学习预测器（预留接口）"""
    
    def __init__(self, model_path: Optional[str] = None):
        self.model = None
        self.scaler = StandardScaler()
        if model_path:
            self.load_model(model_path)
    
    def train(self, X, y):
        """训练模型"""
        # 预留训练接口
        pass
    
    def predict(self, features: Dict) -> Dict:
        """预测故障"""
        # 预留预测接口
        return {
            'ml_risk_score': 0.0,
            'ml_fault_probability': 0.0,
            'ml_fault_type': 'unknown'
        }
    
    def save_model(self, path: str):
        """保存模型"""
        if self.model:
            joblib.dump((self.model, self.scaler), path)
    
    def load_model(self, path: str):
        """加载模型"""
        try:
            self.model, self.scaler = joblib.load(path)
        except:
            pass