requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
scikit-learn==1.3.2
joblib==1.3.2
schedule==1.2.0
openpyxl==3.1.2
openai==1.6.1
anthropic==0.8.1